+++ commons/dimas-core/src/traits.rs
@@ -13,12 +13,13 @@ extern crate std;
 use crate::{
 	enums::{OperationState, TaskSignal},
 	error::Result,
-	message_types::{Message, QueryableMsg}, utils::selector_from,
+	message_types::{Message, QueryableMsg},
+	utils::selector_from,
 };
+use alloc::sync::Arc;
 use core::fmt::Debug;
 #[cfg(feature = "std")]
 use std::prelude::rust_2021::*;
-use alloc::sync::Arc;
 use tokio::sync::mpsc::Sender;
 use zenoh::Session;
 // endregion:	--- modules