+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- filenames of module `liveliness`
+
 ### Fixed
 
 - `Observable`s operation states management
@@ -29,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - example for router (is now in repository [examples](https://github.com/dimas-fw/examples/tree/main/router))
+- no longer needed clippy allows
 
 ## [0.3.0] - 2024-10-03 _Has breaking changes!!_